WS-503: Add Cypress tests
