+++ ws-nextjs-app/cypress/e2e/mediaAssetPage/index.cy.ts
@@ -2,11 +2,11 @@ import { MEDIA_ASSET_PAGE } from '#app/routes/utils/pageTypes';
 import runTestsForPage from '../../support/helpers/runTestsForPage';
 import { testsThatAlwaysRunForAllPages as testsForAllPages } from '../testsForAllPages';
 import { testsThatFollowSmokeTestConfigForAllCanonicalPages as testsForAllCanonicalPages } from '../testsForAllCanonicalPages';
-// import { testsThatFollowSmokeTestConfigForAllAMPPages as testsForAllAMPPages } from '../testsForAllAMPPages';
-// import ampArticleTests from './testsForAMPOnly';
+import { testsThatFollowSmokeTestConfigForAllAMPPages as testsForAllAMPPages } from '../testsForAllAMPPages';
+import ampArticleTests from './testsForAMPOnly';
 import canonicalArticleTests from './testsForCanonicalOnly';
 // import liteTests from '../articlePage/testsForLiteOnly';
-// import getPathWithSuffix from '../../support/helpers/getPathWithSuffix';
+import getPathWithSuffix from '../../support/helpers/getPathWithSuffix';
 
 const canonicalTests = [
   testsForAllPages,
@@ -14,7 +14,7 @@ const canonicalTests = [
   canonicalArticleTests,
 ];
 
-// const ampTests = [testsForAllPages, testsForAllAMPPages, ampArticleTests];
+const ampTests = [testsForAllPages, testsForAllAMPPages, ampArticleTests];
 
 const canonicalSmokeTestSuites = [
   {
@@ -291,14 +291,13 @@ const canonicalTestSuites = Cypress.env('SMOKE')
   ? canonicalSmokeTestSuites
   : canonicalNonSmokeTestSuites;
 
-// SKIPPED: AMP ignored until routes have been fully migrated.
-// const ampTestSuites = canonicalTestSuites.map(testSuite => {
-//   return {
-//     ...testSuite,
-//     path: getPathWithSuffix({ path: testSuite.path, suffix: '.amp' }),
-//     tests: [...ampTests],
-//   };
-// });
+const ampTestSuites = canonicalTestSuites.map(testSuite => {
+  return {
+    ...testSuite,
+    path: getPathWithSuffix({ path: testSuite.path, suffix: '.amp' }),
+    tests: [...ampTests],
+  };
+});
 
 // SKIPPED: LITE ignored until routes have been fully migrated.
 // const liteTestSuites = Cypress.env('SMOKE')
@@ -323,7 +322,7 @@ runTestsForPage({
   testSuites: [
     ...canonicalTestSuites,
     ...tc2CanonicalTestSuites,
-    // ...ampTestSuites,
+    ...ampTestSuites,
     // ...liteTestSuites,
   ],
 });
