+++ ws-nextjs-app/cypress/e2e/photoGalleryPage/index.cy.ts
@@ -1,9 +1,9 @@
-import runTestsForPage from '#nextjs/cypress/support/helpers/runTestsForPage';
-import testsForAllPages from '../testsForAllPages';
+import { PHOTO_GALLERY_PAGE } from '#app/routes/utils/pageTypes';
+import runTestsForPage from '../../support/helpers/runTestsForPage';
+import { testsThatAlwaysRunForAllPages as testsForAllPages } from '../testsForAllPages';
 import { testsThatFollowSmokeTestConfigForAllCanonicalPages as testsForAllCanonicalPages } from '../testsForAllCanonicalPages';
 import { testsThatFollowSmokeTestConfigForAllAMPPages as testsForAllAMPPages } from '../testsForAllAMPPages';
-import { PHOTO_GALLERY_PAGE } from '../../../../src/app/routes/utils/pageTypes';
-import liteArticleTests from '../articles/testsForLiteOnly';
+import liteArticleTests from '../articlePage/testsForLiteOnly';
 
 const tests = [testsForAllPages, testsForAllCanonicalPages];
 
@@ -11,7 +11,7 @@ const canonicalSmokeTestSuites = [
   {
     path: '/pidgin/50913502',
     service: 'pidgin',
-    runforEnv: 'live',
+    runforEnv: ['live'],
     tests,
   },
   {
@@ -23,13 +23,13 @@ const canonicalSmokeTestSuites = [
   {
     path: '/thai/thailand-49950038',
     service: 'thai',
-    runforEnv: 'local',
+    runforEnv: ['local'],
     tests,
   },
   {
     path: '/zhongwen/trad/chinese-news-49065935',
     service: 'zhongwen',
-    runforEnv: 'local',
+    runforEnv: ['local'],
     tests,
   },
 ];
@@ -38,19 +38,19 @@ const canonicalNonSmokeTestSuites = [
   {
     path: '/afaanoromoo/oduu-41217768',
     service: 'afaanoromoo',
-    runforEnv: 'live',
+    runforEnv: ['live'],
     tests,
   },
   {
     path: '/arabic/art-and-culture-38260491',
     service: 'arabic',
-    runforEnv: 'local',
+    runforEnv: ['local'],
     tests,
   },
   {
     path: '/indonesia/indonesia-41635759',
     service: 'indonesia',
-    runforEnv: 'local',
+    runforEnv: ['local'],
     tests,
   },
   {
@@ -62,25 +62,25 @@ const canonicalNonSmokeTestSuites = [
   {
     path: '/mundo/noticias-23147451',
     service: 'mundo',
-    runforEnv: 'test',
+    runforEnv: ['test'],
     tests,
   },
   {
     path: '/persian/magazine-49281981',
     service: 'persian',
-    runforEnv: 'live',
+    runforEnv: ['live'],
     tests,
   },
   {
     path: '/persian/23104784',
     service: 'persian',
-    runforEnv: 'test',
+    runforEnv: ['test'],
     tests,
   },
   {
     path: '/serbian/lat/srbija-46748932',
     service: 'serbian',
-    runforEnv: 'local',
+    runforEnv: ['local'],
     tests,
   },
 ];
@@ -109,5 +109,8 @@ const liteTestSuites = canonicalTestSuites
 
 runTestsForPage({
   pageType: PHOTO_GALLERY_PAGE,
+  headers: {
+    'page-type': 'tc2',
+  },
   testSuites: [...canonicalTestSuites, ...ampTestSuites, ...liteTestSuites],
 });
