+++ ws-nextjs-app/cypress/types.ts
@@ -2,7 +2,8 @@ import { PageTypes, Services, Variants } from '#app/models/types/global';
 
 export type ServiceParametersType = {
   service: Services;
-  variant: Variants;
+  variant?: Variants;
   pageType?: PageTypes & 'errorPage404';
   path?: string;
+  isAmp?: boolean;
 };
